extract album component
