+++ src/components/Album/styles.css
@@ -1,0 +1,17 @@
+
+.album {
+    /* display: inline-block; */
+    height: auto;
+    width: auto;
+    /* float: left; */
+    max-width: 300px;
+    margin: 25px;
+}
+ .album>img {
+    max-width:100%;
+    max-height:100%;
+ }
+
+.info {
+    padding: 1em;
+}
