--- src/components/ArtistPortfolio/styles.css
@@ -1,25 +1,0 @@
-
-.album-container {
-    /* max-width: 100%; */
-    /* display: flex; */
-    /* flex-direction: row wrap; */
-    /* justify-content: center; */
-}
-
-.album {
-    /* display: flex;
-    flex-flow: row wrap;   
-    justify-content: flex-start;
-    margin:  20px; */
-    
-}
-
-.album>img {
-    max-width: 400px;
-    /* width: auto; */
-  
-}
-
-.info {
-    padding: 1em;
-}
