+++ src/pages/index.js
@@ -1,3 +1,4 @@
 export {default as Reviews} from './Reviews'
 export {default as NotFound} from './NotFound'
 export {default as Profile} from './Profile'
+export {default as Albums} from './Albums'
